Split resolver test symbol table helpers
